requirements.txt
torch>=2.0.0
torchvision>=0.15.0
medmnist>=2.1.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
flask>=2.3.0
pillow>=9.0.0
gunicorn>=21.2.0

